Feature/UI improvements (#18)
* Prevent buttons from overlapping

Adds extra padding when a variant is selected to prevent it's toolbar buttons from overlapping the variant selection buttons.

* Improve targeting overview in the UI

Adds target percentage and an icon when the variant has conditions to force show it to the variant buttons in the content block. It helps to see  what the targeting settings are without having to select every single variant.

* Set the default for the optimize block to be enabled

---------

Co-authored-by: Stefan Dürrenberger <sd@pm7.ch>